Implement reference search query DTO
