--- backend-server/src/main/kotlin/de/hsaalen/cmt/services/ServiceItems.kt
@@ -1,9 +1,0 @@
-package de.hsaalen.cmt.services
-
-object ServiceItems {
-
-    suspend fun createItem(title: String){
-
-    }
-
-}
