+++ common/src/commonMain/kotlin/de/hsaalen/cmt/network/client/Requests.kt
@@ -2,6 +2,7 @@ package de.hsaalen.cmt.network.client
 
 import de.hsaalen.cmt.network.RestPaths
 import de.hsaalen.cmt.network.dto.client.ClientLoginDto
+import de.hsaalen.cmt.network.dto.client.ClientReferenceQueryDto
 import de.hsaalen.cmt.network.dto.client.ClientRegisterDto
 import de.hsaalen.cmt.network.dto.server.ServerReferenceListDto
 import de.hsaalen.cmt.network.dto.server.ServerUserInfoDto
@@ -54,7 +55,8 @@ internal object Requests {
     }
 
     /**
-     * Send request to the server for restoring user session. Session can only restored when JWT cookie is still valid.
+     * Send request to the server for restoring user session. Session can only restored when JWT
+     * cookie is still valid.
      */
     suspend fun restore(): ServerUserInfoDto {
         val url = Url("$apiEndpoint/restore")
@@ -64,12 +66,13 @@ internal object Requests {
     }
 
     /**
-     * Provide a list of all related references.
+     * Provide a list of all related references to search query.
      */
-    suspend fun listReferences(): ServerReferenceListDto {
+    suspend fun listReferences(query: ClientReferenceQueryDto): ServerReferenceListDto {
         val url = Url("$apiEndpoint/listReferences")
         return Client.request(url) {
-            method = HttpMethod.Get
+            method = HttpMethod.Post
+            body = query
         }
     }
