+++ common/src/commonMain/kotlin/de/hsaalen/cmt/network/dto/client/ClientReferenceQueryDto.kt
@@ -1,0 +1,11 @@
+package de.hsaalen.cmt.network.dto.client
+
+import kotlinx.serialization.Serializable
+
+/**
+ * Data transfer object to describe a detailed search query to request related references.
+ */
+@Serializable
+data class ClientReferenceQueryDto(
+    val searchName: String? = null
+)
