+++ web-app/src/main/kotlin/de/hsaalen/cmt/pages/OverviewPage.kt
@@ -2,6 +2,7 @@ package de.hsaalen.cmt.pages
 
 import de.hsaalen.cmt.components.ViewReferenceList
 import de.hsaalen.cmt.network.client.Session
+import de.hsaalen.cmt.network.dto.client.ClientReferenceQueryDto
 import de.hsaalen.cmt.network.dto.server.ServerReferenceListDto
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.launch
@@ -17,13 +18,15 @@ class OverviewPage : RComponent<OverviewPage.Props, OverviewPage.State>() {
     }
 
     interface State : RState {
+        var query: ClientReferenceQueryDto
         var dto: ServerReferenceListDto?
     }
 
     override fun State.init() {
+        query = ClientReferenceQueryDto()
         dto = null
         GlobalScope.launch {
-            val received = props.session.listReferences()
+            val received = props.session.listReferences(query)
             setState {
                 dto = received
             }
